xunit, nunit test cases added and src files modified

+++ src/Models/Candle.cs
@@ -3,12 +3,12 @@ namespace CandleRangeAnalyzer.Models
     public class Candle
     {
         public DateTime Time { get; set; }
-        public double Open { get; set; }
-        public double High { get; set; }
-        public double Low { get; set; }
-        public double Close { get; set; }
+        public decimal Open { get; set; }
+        public decimal High { get; set; }
+        public decimal Low { get; set; }
+        public decimal Close { get; set; }
 
-        public double Range => High - Low;
+        public decimal Range => High - Low;
 
         public bool IsBullish => Close > Open;
         public bool IsBearish => Close < Open;
